+++ src/index.ts
@@ -1,68 +1,6 @@
 // Utils
 export { cn } from './lib/utils';
 
-// Business Components
-export { DataTable } from './components/DataTable';
-export { DeleteConfirmDialog } from './components/DeleteConfirmDialog';
-export { TableHeader } from './components/TableHeader';
-export { TablePagination } from './components/TablePagination';
-export { ThemeSwitcher } from './components/ThemeSwitcher';
-export { ThemeSwitcherContent } from './components/ThemeSwitcherContent';
-export { SimplePDFReader } from './components/SimplePDFReader';
-export { PDFReader } from './components/PDFReader';
-export { PDFSidebar } from './components/PDFSidebar';
-export { FileUpload } from './components/FileUpload';
-
-// Types
-export type {
-  Column,
-  DataTableProps,
-  UIComponents,
-} from './components/DataTable';
-export type {
-  DeleteConfirmDialogProps,
-  DialogUIComponents,
-} from './components/DeleteConfirmDialog';
-export type {
-  TableHeaderProps,
-  HeaderUIComponents,
-} from './components/TableHeader';
-export type {
-  TablePaginationProps,
-  PaginationUIComponents,
-} from './components/TablePagination';
-export type {
-  ThemeSwitcherProps,
-  ThemeSwitcherUIComponents,
-  ThemeOption,
-} from './components/ThemeSwitcher';
-export type {
-  ThemeSwitcherContentProps,
-  ThemeSwitcherContentUIComponents,
-} from './components/ThemeSwitcherContent';
-export type {
-  SimplePDFReaderProps,
-  SimplePDFReaderUIComponents,
-} from './components/SimplePDFReader';
-export type {
-  PDFReaderProps,
-  PDFReaderUIComponents,
-} from './components/PDFReader';
-export type {
-  PDFSidebarProps,
-  PDFDocumentProxy,
-  PDFPageProxy,
-  PDFViewport,
-  PDFOutline,
-} from './components/PDFSidebar';
-export type {
-  FileUploadProps,
-  FileUploadUIComponents,
-  FileUploadItem,
-  FileUploadStatus,
-  FileUploadHelpers,
-} from './components/FileUpload';
-
 // Component Types
 export type {
   UIComponent,