+++ src/simple-pdf-reader.ts
@@ -1,0 +1,5 @@
+export { SimplePDFReader } from './components/SimplePDFReader';
+export type {
+  SimplePDFReaderProps,
+  SimplePDFReaderUIComponents,
+} from './components/SimplePDFReader';