+++ src/theme-switcher-content.ts
@@ -1,0 +1,5 @@
+export { ThemeSwitcherContent } from './components/ThemeSwitcherContent';
+export type {
+  ThemeSwitcherContentProps,
+  ThemeSwitcherContentUIComponents,
+} from './components/ThemeSwitcherContent';